FishLight_App: Initial commit
These files come from the "SparkPhone" app that Noah created for me.  I
have just copied it to the new name so that I could have a starting
point for my app updates.  This is a basic app with 2 fields for the
Core ID and Access Token, and 3 fields for the function name, function
argument name, and the command string that gets passed to the function.
It also has 2 buttons, a GET and a POST button to exercise these two
http functions.  I will need to add one for POST to update the Core
firmware at some point.